backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
